For those who want to use bundler to manage their gems (and specifically their gem executibles; e.g., #{Rails.root}/bin/*), it seems it is easiest to modify the environment PATH at the time the cucumber tests are run.  Include the path to browsercms/bin so that the desired executable versions are used in the tests. By not running 'bundle install --binstubs', this path should effectively be ignored.
The Aruba timeout was extended to 180 seconds (3 minutes), because some
of
the bcms calls are taking a little extra time.  Perhaps this is overkill
and
my issue was that I wasn't clearing out the Aruba sandbox after each
test
was run.

Since I depend on #{Rails.root}/bin/*, there is a problem when rake
tasks are
run that the ./bin/bcms* files are modified.  As long as the repo has
the
right versions, a simple git checkout can restore them and the tests
will continue using the correct versions.